// codebases/pipedrive.com(dashboard)/webpack---/client/actions/deals.enum.ts
export enum DealActionTypes {
	MOVE_DEAL_AND_UPDATE = 'MOVE_DEAL_AND_UPDATE',
	MOVE_DEAL = 'MOVE_DEAL',
	ADD_DEAL = 'ADD_DEAL',
	UPDATE_DEAL = 'UPDATE_DEAL',
	DELETE_DEAL = 'DELETE_DEAL',
	UNLOCK_DEAL = 'UNLOCK_DEAL',
	HIDE_DEAL = 'HIDE_DEAL',
	UNHIDE_DEAL = 'UNHIDE_DEAL',
	SET_DEALS_REQUEST = 'SET_DEALS_REQUEST',
	SET_DEALS_SUCCESS = 'SET_DEALS_SUCCESS',
	SET_DEALS_FAILURE = 'SET_DEALS_FAILURE',
	REFRESH_DEAL_FAILURE = 'REFRESH_DEAL_FAILURE',
	RESET_LOADED_DEALS_COUNT = ' RESET_LOADED_DEALS_COUNT',
	SET_DEALS_SUMMARY_REQUEST = 'SET_DEALS_SUMMARY_REQUEST',
	SET_DEALS_SUMMARY_SUCCESS = 'SET_DEALS_SUMMARY_SUCCESS',
	SET_DEALS_SUMMARY_FAILURE = 'SET_DEALS_SUMMARY_FAILURE',
	FORECAST_SET_DEALS_REQUEST = 'FORECAST_SET_DEALS_REQUEST',
	FORECAST_SET_DEALS_SUCCESS = 'FORECAST_SET_DEALS_SUCCESS',
	FORECAST_SET_DEALS_FAILURE = 'FORECAST_SET_DEALS_FAILURE',
}